pyproject.toml
[project]
name = "borderlands-shift-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.1",
    "beautifulsoup4>=4.14.2",
    "discord-py>=2.6.4",
    "python-dotenv>=1.1.1",
]
